Submaster

+++ controllers/endcapsController.js
@@ -1,13 +1,24 @@
 const db = require('../models');
 
 const index = (req, res) => {
-  db.Endcap.find({}).populate('flankA flankB').exec((err, allEndcaps) => {
+  if (!req.session.currentUser){
+    res.status(400);
+    res.send('Not logged In')
+    return;
+  }
+  db.Endcap.find({user: req.session.currentUser._id})
+    .populate('flankA flankB').exec((err, allEndcaps) => {
     if(err) return console.log(err);
     res.json({allEndcaps});
   })
 }
 
 const show = (req, res) => {
+  if (!req.session.currentUser){
+    res.status(400);
+    res.send('Not logged In')
+    return;
+  }
   db.Endcap.findById(req.params.id, (err, foundEndcap) => {
     if(err) console.log(err);
     res.json({foundEndcap});
@@ -15,13 +26,25 @@ const show = (req, res) => {
 }
 
 const create = (req, res) => {
-  db.Endcap.create(req.body, (err, newEndcap) => {
+  if (!req.session.currentUser){
+    res.status(400);
+    res.send('Not logged In')
+    return;
+  }
+  const endcapObj = req.body;
+  endcapObj.user = req.session.currentUser._id
+  db.Endcap.create(endcapObj, (err, newEndcap) => {
     if(err) console.log(err);
     res.json({newEndcap});
   })
 }
 
 const update = (req, res) => {
+  if (!req.session.currentUser){
+    res.status(400);
+    res.send('Not logged In')
+    return;
+  }
   db.Endcap.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -34,6 +57,11 @@ const update = (req, res) => {
 }
 
 const destroy = (req, res) => {
+  if (!req.session.currentUser){
+    res.status(400);
+    res.send('Not logged In')
+    return;
+  }
   db.Endcap.findByIdAndDelete(req.params.id, (err, deletedEndcap) => {
     if(err) console.log(err);
     res.json(deletedEndcap);
